functions/requirements.txt
functions-framework==3.*
firebase-functions==0.1.0
Django==5.2.1
djangorestframework==3.16.0
python-decouple==3.8
psycopg2-binary==2.9.10
supabase==2.16.0
firebase-admin==6.9.0
Pillow==11.2.1
reportlab==4.4.2
qrcode[pil]==8.2
django-cors-headers==4.7.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-widget-tweaks==1.5.0
whitenoise==6.8.2
bleach==6.1.0
python-magic==0.4.27
redis==5.0.1
django-redis==5.4.0
python-dateutil==2.8.2
requests==2.31.0
